backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-magic==0.4.27
openai==1.12.0
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.24.3
Pillow==10.0.0 